Created default shader in Renderer
