+++ TheOneEditor/Renderer3D.cpp
@@ -41,6 +41,9 @@ bool Renderer3D::Start()
 
     CreateCheckersTexture();
 
+    FillDefaultShaders(default_shader_);
+    default_shader_.CompileShader();
+
     return true;
 }
 
@@ -168,6 +171,35 @@ void Renderer3D::CreateCheckersTexture()
     glBindTexture(GL_TEXTURE_2D, 0);
 }
 
+void Renderer3D::FillDefaultShaders(Shader& shader) const {
+    // Simple Vertex Shader
+    shader.vertex_shader =
+        R"(#version 330 core
+		layout (location = 0) in vec3 a_Position;
+		layout (location = 1) in vec2 a_TexCoord;
+		
+		uniform vec4 u_Model;
+		
+		out vec2 v_TexCoord;
+		
+		void main() {
+			gl_Position = u_Model * vec4(a_Position, 1.0f);
+			v_TexCoord = a_TexCoord;
+		})";
+
+    // Simple Fragment Shader
+    shader.fragment_shader =
+        R"(#version 330 core
+		uniform sampler2D u_Texture;
+		
+		in vec2 v_TexCoord;
+		out vec4 color;
+		
+		void main() {
+			color = texture(u_Texture, v_TexCoord);
+		})";
+}
+
 void Renderer3D::CameraInput(double dt)
 {
     if (!app->gui->panelScene->isHovered || app->IsPlaying())
@@ -243,6 +275,6 @@ void Renderer3D::CameraInput(double dt)
 
     if (transform->isDirty())
         transform->updateMatrix();
-        
+
     camera->updateCameraVectors();
 }
